% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moewbgt-package.R
\docType{package}
\name{moewbgt-package}
\alias{moewbgt}
\alias{moewbgt-package}
\title{moewbgt: Access Heat Index (WBGT) and Heat Stroke Alert Data from the Japanese Ministry of the Environment}
\description{
Retrieve and parse heat index (WBGT) forecast and observed values, and heat stroke alert announcements, published by the Heat Illness Prevention Information site of the Japanese Ministry of the Environment (\url{https://www.wbgt.env.go.jp/}). The site serves seasonal CSV downloads and, since fiscal 2026, a WebAPI.
}
\author{
\strong{Maintainer}: Shinya Uryu \email{uryu.shinya@tokushima-u.ac.jp}

Authors:
\itemize{
  \item Shinya Uryu \email{uryu.shinya@tokushima-u.ac.jp}
}

}
\keyword{internal}
