% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wbgt_pref_codes}
\alias{wbgt_pref_codes}
\title{JMA-region codes used by the Ministry of the Environment WebAPI}
\format{
A tibble with 60 rows and four columns:
\describe{
\item{area_cd}{Integer regional code from 1 to 11.}
\item{area_name}{Japanese regional name.}
\item{pref_cd}{Integer prefecture or subprefecture code.}
\item{pref_name}{Japanese prefecture or subprefecture name.}
}
}
\source{
Ministry of the Environment, Heat Illness Prevention Information
Site WebAPI specification, version 1.1, section 2-1,
\url{https://www.wbgt.env.go.jp/man15NH/wbgt_data_api_service_manual.pdf}
}
\usage{
wbgt_pref_codes
}
\description{
A lookup table for the regional and prefecture/subprefecture codes accepted
by \code{getForecastData} and \code{getSurveyData}. These are Japan Meteorological
Agency region codes, not JIS prefecture codes. Hokkaido is divided into
subprefectures and Okinawa uses the combined code 9194.
}
\keyword{datasets}
